# from_table: speedup for sparse and dense data