--- a/Orange/data/util.py
+++ b/Orange/data/util.py
@@ -4,6 +4,7 @@ Data-manipulation utilities.
 import re
 from collections import Counter
 from itertools import chain
+from typing import Callable
 
 import numpy as np
 import bottleneck as bn
@@ -117,32 +118,33 @@ def array_equal(a1, a2):
 def assure_array_dense(a):
     if sp.issparse(a):
         a = a.toarray()
-    return a
+    return np.asarray(a)
 
 
-def assure_array_sparse(a):
+def assure_array_sparse(a, sparse_class: Callable = sp.csc_matrix):
     if not sp.issparse(a):
         # since x can be a list, cast to np.array
         # since x can come from metas with string, cast to float
         a = np.asarray(a).astype(np.float)
-        return sp.csc_matrix(a)
-    return a
+    return sparse_class(a)
 
 
 def assure_column_sparse(a):
-    a = assure_array_sparse(a)
-    # if x of shape (n, ) is passed to csc_matrix constructor,
+    # if x of shape (n, ) is passed to csc_matrix constructor or
+    # sparse matrix with shape (1, n) is passed,
     # the resulting matrix is of shape (1, n) and hence we
     # need to transpose it to make it a column
-    if a.shape[0] == 1:
-        a = a.T
-    return a
+    if a.ndim == 1 or a.shape[0] == 1:
+        # csr matrix becomes csc when transposed
+        return assure_array_sparse(a, sparse_class=sp.csr_matrix).T
+    else:
+        return assure_array_sparse(a, sparse_class=sp.csc_matrix)
 
 
 def assure_column_dense(a):
     a = assure_array_dense(a)
-    # column assignments must be of shape (n,) and not (n, 1)
-    return np.ravel(a)
+    # column assignments must be (n, )
+    return a.reshape(-1)
 
 
 def get_indices(names, name):