--- a/Orange/tests/test_table.py
+++ b/Orange/tests/test_table.py
@@ -1767,6 +1767,36 @@ class CreateTableWithDomainAndTable(TableTests):
         self.assertEqual(back_brown.X.shape, brown.X.shape)
         self.assertEqual(back_brown.metas.shape, brown.metas.shape)
 
+    def test_from_table_shared_compute_value(self):
+        iris = data.Table("iris").to_sparse()
+        d1 = Domain(
+            [
+                ContinuousVariable(
+                    name=at.name,
+                    compute_value=PreprocessSharedComputeValue(
+                        None, PreprocessShared(iris.domain, None)
+                    )
+                )
+                for at in iris.domain.attributes
+            ]
+        )
+
+        new_table = Table.from_table(d1, iris)
+        np.testing.assert_array_equal(
+            new_table.X[:3],
+            [[5.1, 5.1, 5.1, 5.1],
+             [4.9, 4.9, 4.9, 4.9],
+             [4.7, 4.7, 4.7, 4.7]]
+        )
+
+        new_table = Table.from_table(d1, iris, row_indices=[0, 1, 2])
+        np.testing.assert_array_equal(
+            new_table.X[:3],
+            [[5.1, 5.1, 5.1, 5.1],
+             [4.9, 4.9, 4.9, 4.9],
+             [4.7, 4.7, 4.7, 4.7]]
+        )
+
     def assert_table_with_filter_matches(
             self, new_table, old_table,
             rows=..., xcols=..., ycols=..., mcols=...):
@@ -2624,6 +2654,16 @@ class TestTableSparseDense(unittest.TestCase):
         d = self.iris.transform(domain)
         self.assertFalse(sp.issparse(d.X))
 
+        # try with indices that are not Ellipsis
+        d = Table.from_table(domain, self.iris, row_indices=[0, 1, 2])
+        np.testing.assert_array_equal(
+            d.X,
+            [[5.1, 3.5, 1.4, 0.2, 0],
+             [4.9, 3.0, 1.4, 0.2, 0],
+             [4.7, 3.2, 1.3, 0.2, 0]]
+        )
+        self.assertFalse(sp.issparse(d.X))
+
     def test_from_table_add_lots_of_sparse_columns(self):
         n_attrs = len(self.iris.domain.attributes)
 
@@ -2713,7 +2753,8 @@ class PreprocessShared:
         self.callback = callback
 
     def __call__(self, data_):
-        self.callback(data_)
+        if self.callback:
+            self.callback(data_)
         data_.transform(self.domain)
         return True
 
@@ -2726,7 +2767,8 @@ class PreprocessSharedComputeValue(SharedComputeValue):
 
     # pylint: disable=arguments-differ
     def compute(self, data_, shared_data):
-        self.callback(data_)
+        if self.callback:
+            self.callback(data_)
         return data_.X[:, 0]