##### Issue
<!-- E.g. Fixes #1, Closes #2, Resolves #3, etc. -->
<!-- Or a short description, if the issue does not exist. -->
Fixes https://github.com/biolab/orange3/issues/3923

##### Description of changes
Speedup from_table - tested on random data (10000, 3000):
- on dense data: 6s -> 4s
- on sparse data: 350s -> 1.3s

##### Includes
- [X] Code changes
- [X] Tests
- [ ] Documentation