@markotoplak 2020-02-20
@PrimozGodec, before your commit I inserted some speed tests. Please update your branch before making changes. I ran these benchmarks from orange3 folder with:
`python -m unittest -q benchmark.bench_transform`

I get the following results. First, on master:
```
[copy_dense_long] with 5 loops, best of 3:
	min 13.8 msec per loop
	avg 15.2 msec per loop
[copy_dense_square] with 5 loops, best of 3:
	min 52.6 msec per loop
	avg 57.3 msec per loop
[copy_dense_wide] with 2 loops, best of 3:
	min 204 msec per loop
	avg 205 msec per loop
[copy_sparse_long] with 5 loops, best of 3:
	min 2.13 sec per loop
	avg 2.13 sec per loop
[copy_sparse_square] with 5 loops, best of 3:
	min 2.32 sec per loop
	avg 2.33 sec per loop
[copy_sparse_wide] with 2 loops, best of 3:
	min 4.78 sec per loop
	avg 4.79 sec per loop
```

On the new branch:

```
[copy_dense_long] with 5 loops, best of 3:
	min 9.04 msec per loop
	avg 9.07 msec per loop
[copy_dense_square] with 5 loops, best of 3:
	min 26.2 msec per loop
	avg 26.3 msec per loop
[copy_dense_wide] with 2 loops, best of 3:
	min 210 msec per loop
	avg 211 msec per loop
[copy_sparse_long] with 5 loops, best of 3:
	min 56.4 msec per loop
	avg 56.6 msec per loop
[copy_sparse_square] with 5 loops, best of 3:
	min 355 msec per loop
	avg 357 msec per loop
[copy_sparse_wide] with 2 loops, best of 3:
	min 2.7 sec per loop
	avg 2.7 sec per loop
````

Everything improved (`copy_dense_wide` is only few percent slower). While the performance for wide sparse tables did improve by factor 2, it is still seems quite slow. Could you look into it?