@PrimozGodec 2020-02-21
[copy_dense_long] with 5 loops, best of 3:
	min 5.95 msec per loop
	avg 6.05 msec per loop
[copy_dense_square] with 5 loops, best of 3:
	min 17 msec per loop
	avg 18.4 msec per loop
[copy_dense_wide] with 2 loops, best of 3:
	min 126 msec per loop
	avg 126 msec per loop
[copy_sparse_long] with 5 loops, best of 3:
	min 15.7 msec per loop
	avg 17.3 msec per loop
[copy_sparse_square] with 5 loops, best of 3:
	min 131 msec per loop
	avg 138 msec per loop
[copy_sparse_wide] with 2 loops, best of 3:
	min 1.39 sec per loop
	avg 1.41 sec per loop